Initial pass of refactoring gdrive upload
