+++ lib/gdrive/src/shortenUrl.js
@@ -1,10 +1,10 @@
 var request = require('request');
 var tcolorBlue = '\x1b[36m';
 var tcolorReset = '\x1b[0m';
-var config = require('../../config/google-config.json')
+// var config = require('../../../config/google-config.json')
 
-function shortenUrl(longUrl, cb) {
-  request.post('https://www.googleapis.com/urlshortener/v1/url?key=' + config.UrlApiKey,
+function shortenUrl(longUrl, job) {
+  request.post('https://www.googleapis.com/urlshortener/v1/url?key=' + job.config.UrlApiKey,
   { json: { "longUrl": longUrl } },
   function (err, response, body) {
       if (err) {
