+++ lib/gdrive/src/uploadPdf.js
@@ -6,9 +6,9 @@ var shortUrl = '';
 var tcolorBlue = '\x1b[36m';
 var tcolorReset = '\x1b[0m';
 var shortenUrl = require('./shortenUrl');
-var config = require('../../config/google-config.json');
+// var config = require('../../../config/google-config.json');
 
-function uploadPDF(auth, file) {
+function uploadPDF(auth, file, job) {
   var service = google.drive({ version: 'v3', auth: auth });
   var spinner = new Spinner('Uploading PDF.. %s');
   // spinner.setSpinnerString('|/-\\');
@@ -18,7 +18,7 @@ function uploadPDF(auth, file) {
     resource: {
       name: file.name,
       mimeType: 'application/pdf',
-      parents: [config.UploadFolder]
+      parents: [job.config.UploadFolder]
     },
     media: {
       mimeType: 'application/pdf',
@@ -32,7 +32,7 @@ function uploadPDF(auth, file) {
         console.log(tcolorBlue, '\n[API Response]\n', tcolorReset, response);
         var longUrl = 'https://drive.google.com/open?id=' + response.id;
         console.log(tcolorBlue, '\n[Long URL]\n', tcolorReset, longUrl);
-        shortenUrl(longUrl)
+        shortenUrl(longUrl, job)
         spinner.stop();
       }
   });
