+++ lib/gdrive/src/uploadTextFile.js
